add informer for service-controller
